Move download functionality to an assets controller. Log downloads are moved to logs controller. Regenerate rubocop_todo
